+++ package/umt_wasm/src/math/lcm.rs
@@ -1,5 +1,5 @@
 use super::umt_gcd;
 
-pub fn umt_lcm(a: i64, b: i64) -> i64 {
+pub fn umt_lcm(a: i32, b: i32) -> i32 {
     a * b / umt_gcd(a, b)
 }
